# DEV: Speed up system tests by allowing more maximum puma threads