With 0:4 Puma threads:

```
rspec spec/system/social_authentication_spec.rb

Randomized with seed 970
................................................

Finished in 1 minute 55.97 seconds (files took 2.18 seconds to load)
48 examples, 0 failures
```

With 0:20 Puma threads:

```
rspec spec/system/social_authentication_spec.rb

Randomized with seed 14523
................................................

Finished in 1 minute 26.34 seconds (files took 2 seconds to load)
48 examples, 0 failures
```